migrated player list fetch to "PlayersContext"
