+++ screens/Find.jsx
@@ -11,9 +11,9 @@ import * as Location from "expo-location";
 import GradientBackground from "../components/GradientBackground";
 import { colors } from '../styles/colors';
 import { FontAwesome } from '@expo/vector-icons';
+import { usePlayers } from "../contexts/PlayersContext";
 
 export default function Find({ navigation, route }) {
-  const [players, setPlayers] = useState([]);
   const [clickedPlayer, setClickedPlayer] = useState(null);
   const [modalVisible, setModalVisible] = useState(false);
   const [isMapMode, setIsMapMode] = useState(false);
@@ -21,64 +21,13 @@ export default function Find({ navigation, route }) {
     Location.useForegroundPermissions();
   const [hasPermission, setHasPermission] = useState(false);
   const [initialRegion, setInitialRegion] = useState(null);
-
-  useEffect(() => {
-    const q = query(collection(db, "users"));
-
-    const unsubscribe = onSnapshot(q, (querySnapshot) => {
-      if (!querySnapshot.empty) {
-        const players = querySnapshot.docs.map((player) => player.data());
-        const playersWithRank = addRankToPlayers(players);
-        setPlayers(playersWithRank);
-      } else {
-        setPlayers([]);
-      }
-    });
-
-    return () => unsubscribe();
-  }, []);
+  const { players } = usePlayers();
 
   useEffect(() => {
     if (route.params?.isMapMode) {
       setIsMapMode(true);
     }
   }, []);
-
-  function addRankToPlayers(players) {
-    const sortedPlayers = players.sort((a, b) => {
-      return b.score - a.score;
-    });
-
-    if (sortedPlayers.length === 1) {
-      sortedPlayers[0].rank = 1;
-      return sortedPlayers;
-    }
-
-    //add rank field to each player object
-    let rank = 1;
-    let currentUserIndex = 0;
-    sortedPlayers[0].rank = rank;
-    for (let i = 1; i < sortedPlayers.length; i++) {
-      if (sortedPlayers[i].id === auth.currentUser.uid) {
-        currentUserIndex = i;
-      }
-
-      if (sortedPlayers[i].score === sortedPlayers[i - 1].score) {
-        sortedPlayers[i].rank = rank;
-      } else {
-        sortedPlayers[i].rank = i + 1;
-        rank = i + 1;
-      }
-    }
-
-    //insert the current user's player object at the beginning of the array
-    const currentUser = sortedPlayers[currentUserIndex];
-    sortedPlayers.splice(currentUserIndex, 1);
-    sortedPlayers.unshift(currentUser);
-
-    return sortedPlayers;
-
-  }
 
   async function clickHandler(player) {
     setModalVisible(true);
